Create AbstractCacheWrapper
