+++ engine/src/main/java/com/agiletec/aps/system/common/AbstractCacheWrapper.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright 2018-Present Entando Inc. (http://www.entando.com) All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+package com.agiletec.aps.system.common;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.cache.Cache;
+import org.springframework.cache.CacheManager;
+
+/**
+ * @author E.Santoboni
+ */
+public abstract class AbstractCacheWrapper {
+
+	private static final Logger _logger = LoggerFactory.getLogger(AbstractCacheWrapper.class);
+
+	private CacheManager _springCacheManager;
+	
+	protected abstract String getCacheName();
+	
+	protected <T> T get(String name, Class<T> requiredType) {
+		return this.get(this.getCache(), name, requiredType);
+	}
+
+	protected <T> T get(Cache cache, String name, Class<T> requiredType) {
+		Object value = cache.get(name);
+		if (value instanceof Cache.ValueWrapper) {
+			value = ((Cache.ValueWrapper) value).get();
+		}
+		return (T) value;
+	}
+
+	protected Cache getCache() {
+		return this.getSpringCacheManager().getCache(this.getCacheName());
+	}
+
+	protected CacheManager getSpringCacheManager() {
+		return _springCacheManager;
+	}
+
+	public void setSpringCacheManager(CacheManager springCacheManager) {
+		this._springCacheManager = springCacheManager;
+	}
+
+}
